winegcc: Pass through -pthread
While you could get 90% of the way with "-lpthread", supporting -pthread
should be more reliable and is required for build systems like meson
that have special handling for the threads dependency.

Since this is both a compiler (really preprocessor) and linker option,
we can't just add it to is_linker_arg().